Revert "Task 3164"
