+++ l10n_it_asset_management/tests/__init__.py
@@ -1,4 +1,3 @@
 #  License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 from . import test_assets_management
-from . import test_asset_depreciation
