+++ l10n_it_asset_management/views/asset_depreciation.xml
@@ -58,18 +58,6 @@
                                 ],
                             }"
                         />
-                        <field
-                            name="depreciation_account_id"
-                            attrs="{'readonly': [('state', '!=', 'non_depreciated')]}"
-                        />
-                        <field
-                            name="gain_account_id"
-                            attrs="{'readonly': [('state', '!=', 'non_depreciated')]}"
-                        />
-                        <field
-                            name="loss_account_id"
-                            attrs="{'readonly': [('state', '!=', 'non_depreciated')]}"
-                        />
                     </group>
                     <group>
                         <field
