--- setup/l10n_it_asset_management_partner_name_history/odoo/addons/l10n_it_asset_management_partner_name_history
@@ -1,1 +1,0 @@
-../../../../l10n_it_asset_management_partner_name_history
