--- setup/l10n_it_central_journal_partner_name_history/odoo/addons/l10n_it_central_journal_partner_name_history
@@ -1,1 +1,0 @@
-../../../../l10n_it_central_journal_partner_name_history
